Enable building `upb` with Clang.
The version of `upb` used in tensorflow and XLA is incompatible with Clang.
In particular, it generates a warning that the code uses a non-standard
C++ feature. Since this version of `upb` has `-Werror` in its build opts,
the warning breaks the build.

We want to be able to compile PyTorch/XLA with clang, and PyTorch/XLA depends
on `upb`. Therefore we need to make `upb` buildable with Clang.

In this change, we remove `-Werror` from `upb`'s build opts to prevent the
warnings generated by Clang to break the build. In general, we should never
use `-Werror` on code that we don't directly control, as our ability for fixing
the warnings in such code is limited.

PiperOrigin-RevId: 756927870